+++ src/client/pages/dashboard/components/grid_view.vue
