+++ src/client/pages/dashboard/components/list_view.vue
