+++ src/client/pages/dashboard/components/new_project_dialog.vue
