+++ src/client/pages/project/components/beats_viewer.vue
