+++ src/client/pages/project/components/bot_message.vue
