+++ src/client/pages/project/components/presentation_style_selector.vue
