+++ src/client/pages/project/components/product_tabs.vue
