+++ src/client/pages/project/components/prompt_guide.vue
