+++ src/client/pages/project/components/script_editor.vue
